requirements.txt
alembic==1.18.4
blinker==1.9.0
click==8.3.2
croniter==6.2.2
Flask==3.1.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==4.3.0
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
redis==7.4.0
rq==2.7.0
six==1.17.0
SQLAlchemy==2.0.49
typing_extensions==4.15.0
Werkzeug==3.1.8
